shuffle test and tools api
